pyproject.toml
[tool.poetry]
name = "chatgpt-synopsis"
version = "0.1.0"
description = ""
authors = ["RaisCorr <raiscorr@gmail.com>"]
readme = "README.md"
packages = [{include = "chatgpt_synopsis"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
streamlit = "^1.21.0"
openai = "^0.27.4"
replicate = "^0.7.0"
pyaudio = "^0.2.13"
python-dotenv = "^1.0.0"
langchain = "^0.0.135"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
